fix from rebase + golangci-lint
